requirements.txt
discord.py[voice]>=2.3.2
python-dotenv>=1.0.0
yt-dlp>=2024.11.18
PyNaCl>=1.5.0
aiohttp>=3.9.0
groq>=0.4.1
pvporcupine>=3.0.0
pyaudio>=0.2.14
gTTS>=2.5.0
discord-ext-voice-recv
numpy